Fix 0.56.6 native Rust compile regressions
